pyproject.toml
[project]
name = "Chess_teacher"
version = "0.1.0"
description = "Python Boilerplate contains all the boilerplate you need to create a Python package."
readme = "README.md"
authors = [
  {name = "Arnau Biosca Pérez", email = "arnaubi2013@gmail.com"}
]
maintainers = [
  {name = "Arnau Biosca Pérez", email = "arnaubi2013@gmail.com"}
]
classifiers = [
    # TODO
]
license = {text = "MIT"}
dependencies = [
  "pillow>=12.0.0",
  "python-chess>=1.999",
  "rich>=14.2.0",
  "stockfish>=3.28.0",
  "typer",
]
requires-python = ">= 3.10"

[project.optional-dependencies]
test = [
    "coverage",  # testing
    "pytest",  # testing
    "ruff",  # linting
    "ty", # checking types
    "ipdb", # debugging
]

[project.urls]
bugs = "https://github.com/Arnaubiosca15/Chess_teacher/issues"
changelog = "https://github.com/Arnaubiosca15/Chess_teacher/blob/master/changelog.md"
homepage = "https://github.com/Arnaubiosca15/Chess_teacher"

[project.scripts]
Chess_teacher = "Chess_teacher.cli:app"

[tool.ty]
# All rules are enabled as "error" by default; no need to specify unless overriding.
# Example override: relax a rule for the entire project (uncomment if needed).
# rules.TY015 = "warn"  # For invalid-argument-type, warn instead of error.

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "black>=25.9.0",
    "pytest>=8.4.2",
    "ruff>=0.14.3",
]
